Making event list purely integral

+++ anlffr/preproc.py
@@ -84,7 +84,7 @@ def find_blinks(raw, event_id=998, thresh=100e-6, l_freq=0.5, h_freq=10,
     eog_events = np.c_[eog_events, np.zeros(n_events),
                        event_id * np.ones(n_events)]
 
-    return eog_events
+    return np.int64(eog_events)
 
 
 @verbose
@@ -190,8 +190,8 @@ def peak_finder(x0, thresh=None, extrema=1, verbose=None):
             ii += 1  # This is a peak
             # Reset peak finding if we had a peak and the next peak is bigger
             # than the last or the left min was small enough to reset.
-            if found_peak and ((x[ii] > peak_mag[-1])
-                               or (left_min < peak_mag[-1] - thresh)):
+            if found_peak and ((x[ii] > peak_mag[-1]) or
+                               (left_min < peak_mag[-1] - thresh)):
                 temp_mag = min_mag
                 found_peak = False
